# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: IT Asset Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /login:
    post:
      operationId: login
      tags: [auth]
      summary: Employee login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "423":
          description: Account locked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
...
  /admin/users:
    get:
      operationId: getAdminUsers
      tags: [admin]
      summary: Get all users
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
  /admin/users/{id}:
    put:
      operationId: updateAdminUser
      tags: [admin]
      summary: Edit user details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteAdminUser
      tags: [admin]
      summary: Soft delete user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/users/{id}/unlock:
    put:
      operationId: unlockAdminUser
      tags: [admin]
      summary: Unlock locked user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User unlocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/users/{id}/create:
    post:
      operationId: createAdminUser
      tags: [admin]
      summary: Create new employee
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeRequest"
      responses:
        "201":
          description: Employee created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees:
    get:
      operationId: getEmployees
      tags: [employees]
      summary: Get all employees
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Employees list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create employee
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeRequest"
      responses:
        "201":
          description: Employee created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /import-assets:
    post:
      operationId: importAssets
      tags: [import-export]
      summary: Import assets from CSV/XLSX
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
  /export-assets:
    get:
      operationId: exportAssets
      tags: [import-export]
      summary: Export assets to Excel
      security:
        - bearerAuth: []
      parameters:
        - name: format
          in: query
          schema:
            type: string
            enum: [xlsx, csv]
            default: xlsx
      responses:
        "200":
          description: Export file
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    LoginRequest:
      type: object
      properties:
        employeeId:
          type: string
        password:
          type: string
      required:
        - employeeId
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        employee:
          $ref: "#/components/schemas/Employee"
      required:
        - token
        - employee
    Employee:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: string
        name:
          type: string
        email:
          type: string
        department:
          type: string
        role:
          type: string
          enum: [admin, employee]
        isActive:
          type: boolean
        isLocked:
          type: boolean
        failedAttempts:
          type: integer
        lockedAt:
          type: string
          format: date-time
        doj:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - employeeId
        - name
        - role
        - isActive
        - isLocked
        - failedAttempts
    CreateEmployeeRequest:
      type: object
      properties:
        employeeId:
          type: string
        name:
          type: string
        email:
          type: string
        department:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, employee]
        doj:
          type: string
      required:
        - employeeId
        - name
        - role
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        department:
          type: string
        role:
          type: string
          enum: [admin, employee]
        isActive:
          type: boolean
        password:
          type: string
...
    ImportResult:
      type: object
      properties:
        imported:
          type: integer
        skipped:
          type: integer
        errors:
          type: array
          items:
            type: string
      required:
        - imported
        - skipped